+++ .github/workflows/sendNewsletter.yml
@@ -9,6 +9,6 @@ jobs:
       - uses: actions/checkout@v2
       - uses: actions/setup-node@v1
       - run: npm ci
-      - run: npm run sendNewsletter --prefix ./packages/gh-pages
+      - run: npm run sendNewsletter
         env:
           SENDGRID_API_KEY: ${{ secrets.SENDGRID_API_KEY }}
