+++ .github/workflows/tsc.yml
@@ -7,4 +7,4 @@ jobs:
       - uses: actions/checkout@v2
       - uses: actions/setup-node@v1
       - run: npm ci
-      - run: npm run typecheck --prefix ./packages/extension
+      - run: npm run tsc
