+++ .gitignore
@@ -140,3 +140,6 @@ dist
 .vscode-test
 
 # End of https://www.toptal.com/developers/gitignore/api/macos,node
+
+extension.zip
+packages/scripts/screenshots/
