+++ package.json
@@ -1,0 +1,29 @@
+{
+  "private": true,
+  "scripts": {
+    "tsc": "tsc --noEmit",
+    "scrapGames": "node --unhandled-rejections=strict -r ts-node/register ./packages/scripts/scrapGames.ts",
+    "sendNewsletter": "node --unhandled-rejections=strict -r ts-node/register ./packages/scripts/sendNewsletter.ts",
+    "build:extension": "parcel build ./packages/extension/manifest.json -d ./dist/extension",
+    "postbuild:extension": "cd dist/extension && bestzip ../../extension.zip *"
+  },
+  "dependencies": {
+    "@sendgrid/mail": "^7.3.0",
+    "@types/fs-extra": "^9.0.2",
+    "@types/node": "^14.11.10",
+    "@types/puppeteer": "^3.0.2",
+    "@types/react": "^16.9.53",
+    "alpha-sort": "^3.1.0",
+    "bestzip": "^2.1.7",
+    "dom-chef": "^4.2.2",
+    "fs-extra": "^9.0.1",
+    "fuse.js": "^6.4.1",
+    "parcel-bundler": "^1.12.4",
+    "parcel-plugin-web-extension": "^1.6.1",
+    "puppeteer": "^5.3.1",
+    "ts-node": "^9.0.0",
+    "typescript": "^4.0.3",
+    "webext-storage-cache": "^4.1.1",
+    "webextension-polyfill-ts": "^0.20.0"
+  }
+}
