--- packages/extension/.gitignore
@@ -1,1 +1,0 @@
-extension.zip
