--- packages/extension/README.md
@@ -1,3 +1,0 @@
-# extension
-
-This package contains the source of the WebExtension.
