+++ packages/extension/getGame.ts
@@ -1,5 +1,4 @@
 import Fuse from "fuse.js";
-import { APIGame } from "@included-with-xbox-game-pass/types";
 import getGames from "./getGames";
 
 export default async function getGame(name: string): Promise<APIGame | null> {
