+++ packages/extension/getGames.ts
@@ -1,5 +1,4 @@
 import storageCache from "webext-storage-cache";
-import { APIGame } from "@included-with-xbox-game-pass/types";
 
 const API_HOST =
   process.env.NODE_ENV === "production"
