+++ packages/extension/icons/icon96.png
