+++ packages/extension/manifest.json
