--- packages/extension/package.json
@@ -1,23 +1,0 @@
-{
-  "private": true,
-  "scripts": {
-    "typecheck": "tsc --noEmit",
-    "prebuild": "npm run typecheck && rm -rf dist",
-    "build": "parcel build ./src/manifest.json",
-    "postbuild": "cd dist && bestzip ../extension.zip *",
-    "start": "parcel ./src/manifest.json --port 5678"
-  },
-  "devDependencies": {
-    "@included-with-xbox-game-pass/types": "file:../types",
-    "@types/node": "^14.6.0",
-    "@types/react": "^16.9.48",
-    "bestzip": "^2.1.6",
-    "dom-chef": "^4.2.0",
-    "fuse.js": "^6.4.1",
-    "parcel-bundler": "^1.12.4",
-    "parcel-plugin-web-extension": "^1.6.1",
-    "typescript": "^4.0.2",
-    "webext-storage-cache": "^4.0.0",
-    "webextension-polyfill-ts": "^0.19.0"
-  }
-}
