--- packages/gh-pages/.gitignore
@@ -1,1 +1,0 @@
-scripts/screenshots/
