--- packages/gh-pages/README.md
@@ -1,11 +1,0 @@
-# api
-
-This package contains the source of the GitHub page and the "api" (it really is just a json file).
-
-## `static/games.json`
-
-The file `static/games.json` contains the list of games included with the Xbox Game Pass.
-The list is generated by scrapping information from the Xbox Game Pass' website, see `scripts/fetchGame.ts`.
-It's using Puppeteer and prints a few useful information to make sure it was able to fetch all the games' data.
-
-So in order to update the list of games, you have to run `npm run scrapGames`.
