--- packages/gh-pages/package.json
@@ -1,24 +1,0 @@
-{
-  "private": true,
-  "scripts": {
-    "typecheck": "tsc --noEmit",
-    "scrapGames": "node --unhandled-rejections=strict -r ts-node/register ./scripts/scrapGames.ts",
-    "sendNewsletter": "node --unhandled-rejections=strict -r ts-node/register ./scripts/sendNewsletter.ts",
-    "prebuild": "rm -rf dist",
-    "build": "parcel build ./src/index.html --public-url https://gabinaureche.com/included-with-xbox-game-pass/",
-    "start": "parcel ./src/index.html --port 1234"
-  },
-  "devDependencies": {
-    "@included-with-xbox-game-pass/types": "file:../types",
-    "@sendgrid/mail": "^7.2.6",
-    "@types/fs-extra": "^9.0.1",
-    "@types/puppeteer": "^3.0.1",
-    "alpha-sort": "^3.1.0",
-    "fs-extra": "^9.0.1",
-    "parcel-bundler": "^1.12.4",
-    "parcel-plugin-static-files-copy": "^2.5.0",
-    "puppeteer": "^5.2.1",
-    "ts-node": "^8.10.2",
-    "typescript": "^4.0.2"
-  }
-}
