--- packages/gh-pages/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "compilerOptions": {
-    "esModuleInterop": true,
-    "resolveJsonModule": true,
-    "strict": true
-  }
-}
