+++ packages/globals.d.ts
@@ -3,3 +3,16 @@ declare namespace JSX {
   type BaseIntrinsicElement = IntrinsicElements["div"];
   interface IntrinsicAttributes extends BaseIntrinsicElement {}
 }
+
+interface APIGame {
+  id: string;
+  name: string;
+  url: string;
+  image: string;
+  availability: {
+    console: boolean;
+    pc: boolean;
+  };
+  releaseDate: string;
+  addedAt: string;
+}
