--- packages/types/README.md
@@ -1,3 +1,0 @@
-# `types`
-
-This package contains the types shared across the different packages.
