--- packages/types/index.d.ts
@@ -1,12 +1,0 @@
-export interface APIGame {
-  id: string;
-  name: string;
-  url: string;
-  image: string;
-  availability: {
-    console: boolean;
-    pc: boolean;
-  };
-  releaseDate: string;
-  addedAt: string;
-}
