--- packages/types/package.json
@@ -1,7 +1,0 @@
-{
-  "name": "@included-with-xbox-game-pass/types",
-  "version": "0.0.0",
-  "private": true,
-  "main": "",
-  "types": "index.d.ts"
-}
